Ask is now a module, python3 -m ask [script].ask.
