+++ README.md
@@ -66,13 +66,21 @@ if __name__ == '__main__':
 ```
 
 ## Get started
-- Download the `ask.py` file. You can also download `ask-watcher.py` if you want to use live transpilation.
-- Write your code in a `.ask` file
-- Create a python virtual environment (optional): `python3 -m venv venv`
-  - Activate it: `source venv/bin/activate`
-- Download `requirements.txt`, and install the requirements (`pip install -r requirements.txt`)
-- Transpile your code to Flask (Python) with `python3 ask.py [your file].ask`
-- (Optional) If you want to use the watcher: `python3 ask-watcher.py [your file].ask`, the watcher by default uses a 5 second time out between checking your file, you can change this with an optional argument: `python3 ask-watcher.py [your file].ask [number of seconds, ex. 3]`
+- Clone this repository.
+`$ git clone https://github.com/Buscedv/Ask.git`.
+- Make sure that you have Python 3.9 installed. (we recommend that you use [pyenv](https://github.com/pyenv/pyenv)).
+`$ python3 --version`
+
+
+- Install [Poetry](https://python-poetry.org/).
+- Create a virtual environment.
+`$ poetry shell`
+  - (you can exit it with `$ deactivate`).
+- Install depedencies.
+`$ poetry install`
+
+- Create your Ask app in a `.ask` file.
+- Run it: `$ python3 -m ask [path to your app].ask`.
 
 ## Documentation
 You can find the full documentation on [ask.edvard.dev](https://ask.edvard.dev)
